+++ CHANGELOG.md
@@ -4,8 +4,15 @@
 
 ### Features / Enhancements
 
-- Add Disabled Parameter (#24)
-- Add ability to move parameters up and down (#19)
+- Added Disabled Parameter (#24)
+- Added ability to move parameters up and down (#19)
+- Added Split Disabled layout for Input/Output (#27)
+- Added Interpolate Variables (#28)
+- Added Min and Max for Numbers (#29)
+- Added Unit Label (#31)
+- Added Header Parameters (#32)
+- Added Password Input (#33)
+- Subscribed to Refresh Events (#30)
 
 ## 1.0.0 (2022-05-11)